frontend themes: Refactor themes for easier extension
